requirements.txt
streamlit
requests
ffmpeg-python
numpy
openai-whisper